+++ src/Agents/telugu_problem_generator_agent.py
@@ -1,0 +1,29 @@
+##################### Problem Generator #########################
+from .conversable_agent import MyConversableAgent
+from src.Models.llm_config import gpt4_config
+
+class ProblemGeneratorAgent(MyConversableAgent):
+    description = """
+            ProblemGeneratorAgent is a versatile and responsive agent designed to create a wide range of questions to test a StudentAgent's knowledge and skills of Telugu.
+            ProblemGeneratorAgent always provides both Telugu script and romanized Telugu lipi.
+            ProblemGeneratorAgent generates one question at a time.
+            ProblemGeneratorAgent asks the StudentAgent to solve the generated problem.
+            ProblemGeneratorAgent collaborates closely with LevelAdapterAgent to dynamically adjust the complexity of the questions based on StudentAgent's performance.
+            """
+    
+    system_message = """
+            You are ProblemGeneratorAgent, an agent responsible for providing problems to test the StudentAgent's knowledge and skills of Telugu. 
+            ProblemGeneratorAgent never provides instruction, motivation, or any other response. 
+            ProblemGeneratorAgent generates one question at a time.
+            ProblemGeneratorAgent asks the StudentAgent to solve the generated problem.
+            ProblemGeneratorAgent works closely with LevelAdapterAgent, which will monitor the StudentAgent's performance and instruct you when to adjust the difficulty of the questions.             
+            """
+    def __init__(self, **kwargs):        
+        super().__init__(            
+                name="ProblemGeneratorAgent",
+                human_input_mode="NEVER",
+                llm_config=kwargs.pop('llm_config', gpt4_config),
+                system_message=kwargs.pop('system_message', self.system_message),
+                description=kwargs.pop('description', self.description),
+                **kwargs
+            )
